fix: bug in variavel priceCents

+++ RestaurantAPI.Domain/Entities/MenuItem.cs
@@ -7,6 +7,6 @@ namespace RestaurantAPI.Domain.Entities
         [Key]
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public decimal PriceCents { get; set; }
+        public int PriceCents { get; set; }
     }
 }
